+++ meal-ui/src/components/Dropdowns/Dropdown.jsx
@@ -1,5 +1,5 @@
 import { Dropdown } from '@carbon/react';
-import { objectMap, stringifyType } from 'utils';
+import mapToDropdownItems from './mappers';
 
 export function DropdownCustom({ items, onChange, ...additionalProps }) {
 	return (
@@ -9,27 +9,4 @@ export function DropdownCustom({ items, onChange, ...additionalProps }) {
 			{...additionalProps}
 		/>
 	);
-}
-
-function mapToDropdownItems(items) {
-	if (Array.isArray(items)) return mapArray(items);
-	if (typeof items === 'object') return mapObjects(items);
-
-	return [];
-}
-
-function mapArray(items) {
-	return items.map(item => {
-		return {
-			id: item.id ? item.id : 1,
-			label: stringifyType(item.name),
-			...item
-		};
-	});
-}
-
-function mapObjects(items) {
-	return objectMap(items, (key, value) => {
-		return { id: key, label: stringifyType(value), original: value };
-	});
 }
